# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/admin/categories": {
      "get": {
        "tags": [
          "Admin Panel"
        ],
        "summary": "Fetches data for categories page",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "500": {
            "description": "Internal Server Error: Error fetching data for categories page"
          }
        }
      }
    },
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "View all categories",
        "description": "",
        "responses": {
          "200": {
            "description": "OK: Successful operation"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to view category"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Add a new category",
        "description": "You need to execute /auth/login as admin first",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created: Category added successfully",
            "schema": {
              "$ref": "#/definitions/CategoryResponse"
            }
          },
          "400": {
            "description": "Bad Request: Invalid properties"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admin have access"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to add category"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Update a category by ID",
        "description": "You need to execute /auth/login as admin first",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created: Category updated successfully"
          },
          "400": {
            "description": "Bad Request: Invalid properties"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admin have access"
          },
          "404": {
            "description": "Not Found: Category not found"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to update category"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete a category by ID",
        "description": "You need to execute /auth/login as admin first",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK: Category deleted successfully"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admin have access"
          },
          "404": {
            "description": "Not Found: Category not found"
          },
          "409": {
            "description": "Conflict: Cannot delete a category that is in use"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to delete category"
          }
        }
      }
    },
...
    "/checkout/": {
      "get": {
        "tags": [
          "Checkout and Orders"
        ],
        "summary": "View your users own order",
        "description": "You need to execute /auth/login as a regular USER first. You can use the user created when executing /auth/register.<br><br/> <br> Add product to users cart to get successfull result <br/>",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found: No order found"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to view order"
          }
        }
      }
    },
    "/checkout/now": {
      "post": {
        "tags": [
          "Checkout and Orders"
        ],
        "summary": "Checks out users cart",
        "description": "You need to execute /auth/login as a regular USER first. You can use the user created when executing /auth/register. <br><br/> <br> Add product to users cart to get successfull result <br/>",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created: Order created successfully"
          },
          "400": {
            "description": "Bad Request: Invalid properties or values"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "404": {
            "description": "Not found: No products found in cart or membership not found"
          },
          "500": {
            "description": "Internal Server Error: Error occurred while trying to check out cart"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PostCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "New category"
        }
      }
    },
    "UpdateCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Updated category"
        }
      }
    },
...
  }
}